requirements.txt
streamlit
pdfplumber
anthropic
python-dotenv